+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Zippy.java
@@ -3,39 +3,44 @@ package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantm
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
-import net.minecraftforge.common.ForgeMod;
 import net.minecraftforge.event.ItemAttributeModifierEvent;
 
 import java.util.UUID;
 
-public class StepUp extends Enchantment {
+public class Zippy extends Enchantment {
 
     public static final UUID MODIFIER_UUID = UUID.fromString("74e97c20-6a62-482f-b909-e709087b066a");
 
-    public StepUp() {
-        super(Rarity.RARE, EnchantmentCategory.ARMOR_LEGS, new EquipmentSlot[]{EquipmentSlot.LEGS});
+    public Zippy() {
+        super(Rarity.UNCOMMON, EnchantmentCategory.ARMOR_LEGS, new EquipmentSlot[]{EquipmentSlot.LEGS});
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 2;
     }
 
     @Override
     public int getMinCost(int level) {
-        return 20;
+        return 22 * level;
     }
 
     @Override
     public int getMaxCost(int level) {
-        return 40;
+        return this.getMinCost(level) + 22;
     }
 
     public static void applyAttributeModifier(ItemAttributeModifierEvent event) {
         if (event.getSlotType() != EquipmentSlot.LEGS)
             return;
-        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.STEP_UP.get());
+        int lvl = event.getItemStack().getEnchantmentLevel(NewEnchantmentsFeature.ZIPPY.get());
         if (lvl == 0)
             return;
 
-        event.addModifier(ForgeMod.STEP_HEIGHT_ADDITION.get(), new AttributeModifier(MODIFIER_UUID, "Step Up Enchantment Modifier", 0.5d * lvl, AttributeModifier.Operation.ADDITION));
+        event.addModifier(Attributes.MOVEMENT_SPEED, new AttributeModifier(MODIFIER_UUID, "Zippy Enchantment Modifier", 0.1d * lvl, AttributeModifier.Operation.MULTIPLY_BASE));
     }
 
 }